Moving wait methods to utils directory
